Fix validation objectID and other minor bugs
