+++ connectdb.js
@@ -1,0 +1,14 @@
+const mongoose = require('mongoose');
+
+const connectDB = () => {
+  mongoose.connect('mongodb://localhost:27017/mestodb', {
+    useNewUrlParser: true,
+    useCreateIndex: true,
+    useFindAndModify: false,
+    useUnifiedTopology: true,
+  });
+};
+
+module.exports = {
+  connectDB,
+};
