+++ routes/cards.js
@@ -3,9 +3,10 @@ const router = require('express').Router({ mergeParams: true });
 const {
   getAllCards, postCard, deleteCard,
 } = require('../controllers/cards');
+const { cardValidator, mongooseObjectIdValidator } = require('../middlewares/valid_celebrate');
 
 router.get('/cards', getAllCards);
-router.post('/cards', postCard);
-router.delete('/cards/:id', deleteCard);
+router.post('/cards', cardValidator, postCard);
+router.delete('/cards/:id', mongooseObjectIdValidator, deleteCard);
 
 module.exports = router;
