+++ routes/users.js
@@ -4,8 +4,9 @@ const {
   getAllUsers, getUser,
 } = require('../controllers/users');
 
+const { mongooseObjectIdValidator } = require('../middlewares/valid_celebrate');
 
-router.get('/users/:id', getUser);
+router.get('/users/:id', mongooseObjectIdValidator, getUser);
 router.get('/users', getAllUsers);
